MULE-11737: Remove Transient registry
