+++ core/src/main/java/org/mule/runtime/core/internal/registry/MuleContextProcessor.java
@@ -20,12 +20,13 @@ import org.mule.runtime.core.privileged.registry.InjectProcessor;
 @Deprecated
 public class MuleContextProcessor implements InjectProcessor {
 
-  private MuleContext context;
+  private final MuleContext context;
 
   public MuleContextProcessor(MuleContext context) {
     this.context = context;
   }
 
+  @Override
   public Object process(Object object) {
     if (object instanceof MuleContextAware) {
       ((MuleContextAware) object).setMuleContext(context);
