+++ core/src/main/java/org/mule/runtime/core/internal/registry/MuleRegistryHelper.java
@@ -18,7 +18,6 @@ import org.mule.runtime.core.api.construct.FlowConstruct;
 import org.mule.runtime.core.privileged.registry.RegistrationException;
 
 import java.util.Collection;
-import java.util.HashMap;
 import java.util.Map;
 
 /**
@@ -32,8 +31,6 @@ public class MuleRegistryHelper implements MuleRegistry {
   private final Registry registry;
 
   private final MuleContext muleContext;
-
-  private final Map<Object, Object> postProcessedObjects = new HashMap<>();
 
   public MuleRegistryHelper(Registry registry, MuleContext muleContext) {
     this.registry = registry;
@@ -101,7 +98,7 @@ public class MuleRegistryHelper implements MuleRegistry {
    */
   @Override
   public void registerFlowConstruct(FlowConstruct flowConstruct) throws MuleException {
-    registry.registerObject(getName(flowConstruct), flowConstruct, FlowConstruct.class);
+    registry.registerObject(getName(flowConstruct), flowConstruct);
   }
 
   /**
@@ -206,15 +203,6 @@ public class MuleRegistryHelper implements MuleRegistry {
   @Override
   public void registerObject(String key, Object value, Object metadata) throws RegistrationException {
     registry.registerObject(key, value, metadata);
-
-    postObjectRegistrationActions(value);
-  }
-
-  public void postObjectRegistrationActions(Object value) {
-    // TODO MULE-10238 - Remove this check once SimpleRegistry gets removed
-    if (!postProcessedObjects.containsKey(value)) {
-      postProcessedObjects.put(value, value);
-    }
   }
 
   /**
@@ -223,8 +211,6 @@ public class MuleRegistryHelper implements MuleRegistry {
   @Override
   public void registerObject(String key, Object value) throws RegistrationException {
     registry.registerObject(key, value);
-
-    postObjectRegistrationActions(value);
   }
 
   /**
@@ -233,18 +219,6 @@ public class MuleRegistryHelper implements MuleRegistry {
   @Override
   public void registerObjects(Map objects) throws RegistrationException {
     registry.registerObjects(objects);
-
-    for (Object value : objects.values()) {
-      postObjectRegistrationActions(value);
-    }
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public Object unregisterObject(String key, Object metadata) throws RegistrationException {
-    return registry.unregisterObject(key, metadata);
   }
 
   /**
