+++ modules/spring-config/src/main/java/org/mule/runtime/config/internal/processor/PostRegistrationActionsPostProcessor.java
@@ -83,7 +83,6 @@ public class PostRegistrationActionsPostProcessor implements BeanPostProcessor {
         return bean;
       }
       seenBeanNames.add(beanName);
-      registryHelper.postObjectRegistrationActions(object);
     }
     return bean;
   }
