+++ tests/unit/src/main/java/org/mule/runtime/core/internal/config/builders/SimpleRegistryBootstrap.java
@@ -24,8 +24,6 @@ import org.mule.runtime.core.internal.config.bootstrap.ObjectBootstrapProperty;
 import org.mule.runtime.core.internal.config.bootstrap.TransformerBootstrapProperty;
 import org.mule.runtime.core.internal.context.MuleContextWithRegistry;
 import org.mule.runtime.core.internal.registry.SimpleRegistry;
-import org.mule.runtime.core.internal.util.StreamCloser;
-import org.mule.runtime.core.privileged.registry.ObjectProcessor;
 import org.mule.runtime.core.privileged.registry.RegistrationException;
 
 /**
@@ -73,16 +71,11 @@ public class SimpleRegistryBootstrap extends AbstractRegistryBootstrap {
   @Override
   protected void doRegisterObject(ObjectBootstrapProperty bootstrapProperty) throws Exception {
     Object value = bootstrapProperty.getService().instantiateClass(bootstrapProperty.getClassName());
-    Class<?> meta = Object.class;
 
-    if (value instanceof ObjectProcessor) {
-      meta = ObjectProcessor.class;
-    } else if (value instanceof StreamCloser) {
-      meta = StreamCloser.class;
-    } else if (value instanceof BootstrapObjectFactory) {
+    if (value instanceof BootstrapObjectFactory) {
       setMuleContextIfNeeded(value, muleContext);
       value = ((BootstrapObjectFactory) value).create();
     }
-    ((MuleContextWithRegistry) muleContext).getRegistry().registerObject(bootstrapProperty.getKey(), value, meta);
+    ((MuleContextWithRegistry) muleContext).getRegistry().registerObject(bootstrapProperty.getKey(), value);
   }
 }
